backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiohttp-jinja2==1.6
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.74.1
anyio==4.11.0
asgiref==3.9.2
attrs==25.4.0
cachetools==6.2.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
coverage==7.11.3
diskcache==5.6.3
distro==1.9.0
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.17.0
faiss-cpu==1.12.0
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
linkify-it-py==2.0.3
llama_cpp_python==0.3.16
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdit-py-plugins==0.5.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==2.8.1
ordered-set==4.1.0
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pytest==9.0.0
pytest-cov==7.0.0
pytest-django==4.11.1
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
textual==6.4.0
textual-dev==1.8.0
textual-serve==1.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
triton==3.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uc-micro-py==1.0.3
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.22.0
